Only start uploader when all files have been setup
Fixes #84

When uploading muliple files which upload-setup depended on an
async operation (e.g. when obtaining the upload url from a
backend-api) the uploader component contained a race condition
where it assumed that the first file in the plupload uploader
`FilesAdded` callback would also be the first to start uploading.

If that wasn't the case and due to the async nature of obtaining
relevant upload-information one of the files that was queued for
a later upload in plupload was to call its file#upload-method
first the upload of the first file in the queue would fail as it
hadn't completed its setup completely yet.

By making sure that the uploader will only start when all of the
files that should be uploaded have been setup correctly we make
sure that this race condition is no longer an issue.